Updated Brutus.jl to use GPUCompiler.cachedcompilation. Julia tests passing.
