+++ Brutus/Project.toml
@@ -3,5 +3,8 @@ uuid = "44ccd279-0a44-4492-af09-0e34b2907bcc"
 authors = ["Valentin Churavy <v.churavy@gmail.com>", "Leon Shen <leon@yhls.org>"]
 version = "0.1.0"
 
+[deps]
+GPUCompiler = "61eb1bfa-7361-4325-ad38-22787b887f55"
+
 [compat]
 julia = "1.5"
