+++ Brutus/src/intrinsics.jl
@@ -1,0 +1,5 @@
+abstract type BrutusIntrinsic <: Function end
+
+struct Matmul! <: BrutusIntrinsic end
+@noinline (::Matmul!)(A, B, C) = Base.inferencebarrier(nothing)::Nothing
+const matmul! = Matmul!()
